update dplyr checking

+++ src/geos.cpp
@@ -393,7 +393,7 @@ Rcpp::List CPL_geos_op(std::string op, Rcpp::List sfc,
 	std::vector<GEOSGeom> out(sfc.length());
 
 	if (op == "buffer") {
-		if (bufferDist.size() != g.size())
+		if (bufferDist.size() != (int) g.size())
 			throw std::invalid_argument("invalid dist argument"); // #nocov
 		for (size_t i = 0; i < g.size(); i++)
 			out[i] = chkNULL(GEOSBuffer_r(hGEOSCtxt, g[i], bufferDist[i], nQuadSegs));
